+++ crypto/legacytx/access_list.go
@@ -1,7 +1,7 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
 
-package ostx
+package legacytx
 
 import (
 	"github.com/ethereum/go-ethereum/common"
