+++ crypto/legacytx/legacy_tx.go
@@ -1,7 +1,7 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
 
-package ostx
+package legacytx
 
 import (
 	"math/big"
